+++ resources/examples/helloWorld.js
@@ -5,3 +5,4 @@ vscode.window.showInformationMessage(
   `Hello, World! This is ${basename(macros.macro.uri.fsPath)}.`,
   { modal: true },
 );
+macros.log.info('Greeted the world');