+++ src/extensionContext.ts
@@ -6,12 +6,14 @@ export class ExtensionContext implements vscode.Disposable {
   private _mruMacro?: vscode.Uri;
   public readonly extensionContext: vscode.ExtensionContext;
   public readonly libraryManager: MacroLibraryManager;
+  public readonly log: vscode.LogOutputChannel;
   public readonly runnerManager: MacroRunnerManager;
 
   constructor(extensionContext: vscode.ExtensionContext) {
     this.extensionContext = extensionContext;
     this.libraryManager = new MacroLibraryManager();
-    this.runnerManager = new MacroRunnerManager();
+    this.log = vscode.window.createOutputChannel('Macros', { log: true });
+    this.runnerManager = new MacroRunnerManager(this);
     this.runnerManager.onRun(({ macro: { uri } }) => {
       this._mruMacro = uri;
       vscode.commands.executeCommand('setContext', 'macros:mruSet', true);
@@ -21,6 +23,9 @@ export class ExtensionContext implements vscode.Disposable {
   dispose() {
     this.libraryManager.dispose();
     this.runnerManager.dispose();
+
+    // Dispose logger last
+    this.log.dispose();
   }
 
   public get mruMacro(): vscode.Uri | undefined {